datashack/requirements.txt
requests==2.28.1
pandas==1.5.2
pyyaml
dataclasses_avroschema
streamlit==1.15.2
PyHive==0.6.5
SQLAlchemy==1.4.45
thrift==0.16.0
thrift-sasl==0.4.3
# sasl==0.3.1
confluent-kafka==1.9.2
python-terraform==0.10.1
streamlit-option-menu==0.3.2
flask
retrying==1.3.4